chore: split schema
